requirements.txt
watchdog==6.0.0
pypdf==6.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1